+++ docs/changelog.rst
@@ -9,15 +9,16 @@ Changelog
 Added
 ~~~~~
 
-- Possibility to truncate data in the database instead of the DB re-creation.
-  ``truncate`` method in DB backends and ``--truncate`` option for Django integration. `#48`_
+- Possibility to truncate data in the database instead of the DB re-creation. ``truncate`` method in DB backends. `#48`_
 - Command Line Interface. `#6`_
 
 Changed
 ~~~~~~~
 
-- Do not recreate the DB if it the schema is absent in the dump. `#39`_
+- Do not try to load schema if it is absent in the dump. `#39`_
 - ``xdump.sqlite.SQLiteBackend`` now accepts only ``dbname`` and ``verbosity``. `#52`_
+- Now Django's ``xload`` command doesn't recreate the DB, but provides a non required option ``--cleanup-method`` for
+  choosing how to clean-up the database before the dump loading. `#53`_
 
 `0.5.0`_ - 2018-08-02
 ---------------------
@@ -106,6 +107,7 @@ Fixed
 .. _0.1.2: https://github.com/Stranger6667/xdump/compare/0.1.1...0.1.2
 .. _0.1.1: https://github.com/Stranger6667/xdump/compare/0.1.0...0.1.1
 
+.. _#53: https://github.com/Stranger6667/xdump/issues/53
 .. _#52: https://github.com/Stranger6667/xdump/issues/52
 .. _#48: https://github.com/Stranger6667/xdump/issues/48
 .. _#45: https://github.com/Stranger6667/xdump/issues/45